Popover sticky (in-progress), refactoring BarPlotChart into multiple files as was starting to become quite long (1.5k lines).
