+++ src/encoded/static/components/app.js
@@ -424,6 +424,7 @@ var App = React.createClass({
             if (prevState.session !== this.state.session && ChartDataController.isInitialized()){
                 setTimeout(function(){
                     // Delay 5s.
+                    console.log("SYNCING CHART DATA");
                     ChartDataController.sync();
                 }, 5000);   
             }
@@ -876,22 +877,6 @@ var App = React.createClass({
         }
     },
 
-    /**
-     * Use this function, e.g. via passing down through props, to update the ChartDetailCursor component.
-     * 
-     * @public
-     * @instance
-     * @param {Object} d - The state to pass and set on ChartDetailCursor.
-     */
-    updateChartDetailCursor : _.debounce(function(d){
-        return (
-            this.refs &&
-            this.refs.detailCursor &&
-            this.refs.detailCursor.update &&
-            this.refs.detailCursor.update(d)
-        ) || null;
-    }, 100),
-
     render: function() {
         console.log('render app');
         var context = this.props.context;
@@ -1124,11 +1109,9 @@ var App = React.createClass({
                         node.style.top = null;
                     }} />
                     <ChartDetailCursor
-                        containingElement={(this.refs && this.refs.mosaicContainer) || null}
                         verticalAlign="center" /* cursor position */
                         visibilityMargin={{ left : -10, right : -10, bottom : -50, top: -18 }}
                         //debugStyle /* -- keep this Component always visible so we can style it */
-                        ref="detailCursor"
                     />
                 </body>
             </html>
